Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 //@Component Annotation que registra a classe como um component para o spring(permitindo o @Autowired do service).
@@ -35,7 +37,13 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);	// Metodo para deletar usuario por id.
+		try {
+			if(!repository.existsById(id)) throw new ResourceNotFoundException(id); 
+			repository.deleteById(id);	// Metodo para deletar usuario por id.
+		}
+		catch(DataIntegrityViolationException e){
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
